Add joint gpu and cpu modes
