+++ src/main/java/com/firstperson/gpu/GLBuffer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Lotto <https://github.com/devLotto>
+ * Copyright (c) 2021, Adam <Adam@sigterm.info>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,27 +22,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.firstperson.config;
+package com.firstperson.gpu;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum AntiAliasingMode
+class GLBuffer
 {
-	DISABLED("Disabled", 0),
-	MSAA_2("MSAA x2", 2),
-	MSAA_4("MSAA x4", 4),
-	MSAA_8("MSAA x8", 8),
-	MSAA_16("MSAA x16", 16);
+	String name;
+	int glBufferId = -1;
+	int size = -1;
+	long clBuffer = -1;
 
-	private final String name;
-	private final int samples;
-
-	@Override
-	public String toString()
+	GLBuffer(String name)
 	{
-		return name;
+		this.name = name;
 	}
 }
