+++ src/main/java/com/firstperson/gpu/Shader.java
@@ -23,9 +23,9 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.firstperson;
+package com.firstperson.gpu;
 
-import com.firstperson.template.Template;
+import net.runelite.client.plugins.gpu.template.Template;
 import com.google.common.annotations.VisibleForTesting;
 import java.util.ArrayList;
 import java.util.List;
