+++ src/main/resources/com/firstperson/gpu/.clang-format
