+++ src/main/resources/com/firstperson/gpu/cl_types.cl
