+++ src/main/resources/com/firstperson/gpu/colorblind.glsl
