+++ src/main/resources/com/firstperson/gpu/common.cl
