+++ src/main/resources/com/firstperson/gpu/comp.cl
