+++ src/main/resources/com/firstperson/gpu/comp_common.glsl
