+++ src/main/resources/com/firstperson/gpu/frag.glsl
