+++ src/main/resources/com/firstperson/gpu/fragui.glsl
