+++ src/main/resources/com/firstperson/gpu/geom.glsl
