+++ src/main/resources/com/firstperson/gpu/hsl_to_rgb.glsl
