+++ src/main/resources/com/firstperson/gpu/priority_render.cl
