+++ src/main/resources/com/firstperson/gpu/regions/regions.txt
