+++ src/main/resources/com/firstperson/gpu/scale/bicubic.glsl
