+++ src/main/resources/com/firstperson/gpu/scale/xbr_lv2_frag.glsl
