+++ src/main/resources/com/firstperson/gpu/vert.glsl
